+++ src/database/tests/QueryBuilderTest.php
@@ -2192,14 +2192,14 @@ class QueryBuilderTest extends TestCase
     {
         $builder = $this->getMySqlBuilder();
         $builder->select('*')->from('users')->where('items->available', '=', true);
-        $this->assertEquals('select * from `users` where json_unquote(json_extract(`items`, \'$."available"\')) = true', $builder->toSql());
+        $this->assertEquals('select * from `users` where json_extract(`items`, \'$."available"\') = true', $builder->toSql());
     }
 
     public function testMySqlWrappingJsonWithBooleanAndIntegerThatLooksLikeOne()
     {
         $builder = $this->getMySqlBuilder();
         $builder->select('*')->from('users')->where('items->available', '=', true)->where('items->active', '=', false)->where('items->number_available', '=', 0);
-        $this->assertEquals('select * from `users` where json_unquote(json_extract(`items`, \'$."available"\')) = true and json_unquote(json_extract(`items`, \'$."active"\')) = false and json_unquote(json_extract(`items`, \'$."number_available"\')) = ?', $builder->toSql());
+        $this->assertEquals('select * from `users` where json_extract(`items`, \'$."available"\') = true and json_extract(`items`, \'$."active"\') = false and json_unquote(json_extract(`items`, \'$."number_available"\')) = ?', $builder->toSql());
     }
 
     public function testMySqlWrappingJson()
@@ -2876,6 +2876,49 @@ class QueryBuilderTest extends TestCase
         ]), $result);
     }
 
+    public function testPaginateWithTotalOverride()
+    {
+        $page = 1;
+        $total = 10;
+        $builder = $this->getMockQueryBuilder();
+        $path = 'http://foo.bar?page=3';
+
+        $results = collect();
+
+        $builder->shouldReceive('get')->times(2)->andReturn($results);
+        $builder->shouldReceive('getCountForPagination')->andReturn($total);
+
+        Context::set('path', $path);
+        Context::set('page', $page);
+        if (! defined('BASE_PATH')) {
+            define('BASE_PATH', __DIR__);
+        }
+
+        $container = Mockery::mock(Container::class);
+        $container->shouldReceive('make')->times(2)->andReturnUsing(function ($interface, $args) {
+            /** @var Collection $items */
+            $items = $args['items'];
+            return new LengthAwarePaginator(collect($items), $args['total'], $args['perPage'], $args['currentPage'], $args['options']);
+        });
+        ApplicationContext::setContainer($container);
+
+        Paginator::currentPageResolver(function () {
+            return Context::get('page');
+        });
+
+        Paginator::currentPathResolver(function () {
+            return Context::get('path');
+        });
+
+        $result = $builder->paginate(total: 10);
+
+        $this->assertEquals(10, $result->total());
+
+        $result = $builder->paginate(total: fn () => 10);
+
+        $this->assertEquals(10, $result->total());
+    }
+
     public function testWhereRowValues()
     {
         $builder = $this->getBuilder();
@@ -2997,6 +3040,15 @@ class QueryBuilderTest extends TestCase
             ->where('last_seen_at', '>', '1520652582');
         $this->assertEquals('select * from (select max(last_seen_at) as last_seen_at from "sessions") as "last_seen_at" where "last_seen_at" > ?', $builder->toSql());
         $this->assertEquals(['1520652582'], $builder->getBindings());
+    }
+
+    public function testClone()
+    {
+        $builder = $this->getBuilder();
+        $clone = $builder->clone();
+        $this->assertEquals($builder->toSql(), $clone->toSql());
+        $this->assertEquals($builder->getBindings(), $clone->getBindings());
+        $this->assertNotSame($builder, $clone);
     }
 
     protected function getBuilderWithProcessor(): Builder