+++ src/paginator/src/Paginator.php
@@ -34,6 +34,8 @@ class Paginator extends AbstractPaginator implements Arrayable, Countable, Itera
      */
     public function __construct($items, int $perPage, ?int $currentPage = null, array $options = [])
     {
+        $this->options = $options;
+
         foreach ($options as $key => $value) {
             $this->{$key} = $value;
         }