+++ src/swagger/src/Command/GenCommand.php
@@ -11,51 +11,34 @@ declare(strict_types=1);
  */
 namespace Hyperf\Swagger\Command;
 
-use Hyperf\Command\Command;
-use OpenApi\Analysis;
-use Symfony\Component\Console\Input\InputOption;
+use Hyperf\Command\Command as HyperfCommand;
+use Hyperf\Contract\ConfigInterface;
+use Hyperf\Swagger\Generator;
+use Psr\Container\ContainerInterface;
 
-use function OpenApi\scan;
-
-use const OpenApi\UNDEFINED;
-
-class GenCommand extends Command
+class GenCommand extends HyperfCommand
 {
-    protected ?string $name = 'swagger:gen';
+    public function __construct(protected ContainerInterface $container)
+    {
+        parent::__construct('gen:swagger');
+    }
+
+    public function configure()
+    {
+        parent::configure();
+        $this->setDescription('Generate swagger json file.');
+    }
 
     public function handle()
     {
-        $path = $this->input->getOption('path');
-        $outputDir = $this->input->getOption('output');
-        $format = $this->input->getOption('format');
-        Analysis::registerProcessor(function (Analysis $analysis) {
-            foreach ($analysis->openapi->paths ?? [] as $path) {
-                if ($path->path !== UNDEFINED) {
-                    continue;
-                }
-                switch ($path) {
-                    case isset($path->get):
-                        $operationId = $path->get->operationId;
-                        if (strpos($operationId, '::') !== false) {
-                            [$controller, $action] = explode('::', $operationId);
-                            // @TODO Retrieve the path according to controller and action name, and then set the path to $path->path.
-                        }
-                        break;
-                }
-            }
-        });
-        $scanner = scan($path);
-        $destnation = $outputDir . 'openapi.' . $format;
-        $scanner->saveAs($destnation);
-        $this->info(sprintf('[INFO] Written to %s successfully.', $destnation));
-    }
+        $config = $this->container->get(ConfigInterface::class);
 
-    protected function getOptions(): array
-    {
-        return [
-            ['path', 'p', InputOption::VALUE_OPTIONAL, 'The path that needs scan.', 'app/'],
-            ['output', 'o', InputOption::VALUE_OPTIONAL, 'Path to store the generated documentation.', './'],
-            ['format', 'f', InputOption::VALUE_OPTIONAL, 'The format of the generated documentation, supports yaml and json.', 'yaml'],
-        ];
+        // are already generated in the listener if Swagger is enabled and automatically generated.
+        if (! ($config->get('swagger.enable', false) && $config->get('swagger.auto_generate', false))) {
+            $generator = $this->container->get(Generator::class);
+            $generator->generate();
+        }
+
+        $this->output->writeln('Generate swagger json success.');
     }
 }